LPS-159531 Create useModalContext to reuse Component to other Pages
