+++ modules/apps/site-initializer/site-initializer-testray/extra/remote-app/src/components/Sidebar/useSidebarActions.tsx
@@ -12,12 +12,10 @@
  * details.
  */
 
-import ClayButton from '@clayui/button';
-import {Context} from '@clayui/modal';
-import {Size} from '@clayui/modal/lib/types';
-import React, {ReactElement, useContext} from 'react';
+import React from 'react';
 
 import useFormModal from '../../hooks/useFormModal';
+import useModalContext from '../../hooks/useModalContext';
 import i18n from '../../i18n';
 import ProjectModal from '../../pages/Project';
 import CaseTypeModal from '../../pages/Standalone/CaseType/CaseTypeModal';
@@ -25,45 +23,9 @@ import FactorCategoryModal from '../../pages/Standalone/FactorCategory/FactorCat
 import OptionsModal from '../../pages/Standalone/FactorOptions/FactorOptionsModal';
 import {LIFERAY_URLS} from '../../services/liferay';
 
-interface ModalOptions {
-	body: ReactElement;
-	footer?: ReactElement;
-	size: Size;
-	title: string;
-}
-
 const useSidebarActions = () => {
 	const {modal} = useFormModal();
-	const [state, dispatch] = useContext(Context);
-
-	const onOpenModal = ({body, footer, size, title}: ModalOptions) => {
-		dispatch({
-			payload: {
-				body,
-				footer: footer
-					? [
-							undefined,
-							undefined,
-							<ClayButton.Group key={3} spaced>
-								<ClayButton
-									displayType="secondary"
-									onClick={state.onClose}
-								>
-									Cancel
-								</ClayButton>
-
-								<ClayButton key={4} onClick={state.onClose}>
-									Save
-								</ClayButton>
-							</ClayButton.Group>,
-					  ]
-					: [],
-				header: title,
-				size,
-			},
-			type: 1,
-		});
-	};
+	const {onOpenModal} = useModalContext();
 
 	const MANAGE_DROPDOWN = [
 		{
